+++ src/string/contains.js
@@ -3,11 +3,10 @@ define(['../lang/toString'], function(toString) {
     /**
      * Searches for a given substring
      */
-    function contains(str, substring){
+    function contains(str, substring, fromIndex){
         str = toString(str);
         substring = toString(substring);
-
-        return str.indexOf(substring) !== -1;
+        return str.indexOf(substring, fromIndex) !== -1;
     }
 
     return contains;